Switch to Cygwin for Windows build environment
MinGW-w64 MSYS doesn't have mingw-get.  Cygwin, on the other hand, has
a much more complete package set (so we don't need to build pkg-config
or nasm) and nice packages of MinGW-w64 32- and 64-bit cross-compilers.

Installing Cygwin packages requires the original setup.exe, which is not
installed in a fixed location, so we need a special setup subcommand.