Disable -ftree-vectorize
Toby Cornish reported crashes in OpenSlide Python, reproducible on
32-bit Windows:

http://lists.andrew.cmu.edu/pipermail/openslide-users/2015-May/001060.html

These seem to go away if either -msse2 -mfpmath=sse or -ftree-vectorize
is disabled, but if SSE2 is disabled and -ftree-vectorize is not,
F21 mingw32-gcc (4.9.2) hits an ICE when building pixman.  Disable
-ftree-vectorize under the assumption that it is the more invasive/
less tested option.  (However, it is still necessary to filter out
-msse2 when building OpenJPEG.)