Drop _FORTIFY_SOURCE; conditionally link with libssp
According to this thread, -D_FORTIFY_SOURCE=2 historically hasn't done
anything:

    https://sourceforge.net/p/mingw-w64/mailman/mingw-w64-public/thread/8f541ca2-fb82-4f89-929c-4bf994da6c83%40126.com/

But it does now, and uses symbols from libssp.  However, we can't
actually enable _FORTIFY_SOURCE (or -fstack-protector-strong) because
that causes the libssp __stack_chk_guard to be initialized from strong
entropy at DLL load time, causing CryptAcquireContext() to deadlock on
the DLL loader lock.  It's not clear how this ever worked properly, but
I haven't debugged in detail; maybe it only happens with a particular
initialization chain in one of our libraries.  For now, disable
_FORTIFY_SOURCE to work around this.  Since that macro didn't do
anything before, we're no worse off.

Further complicating things, the MinGW runtime in Fedora is built with
_FORTIFY_SOURCE, though apparently not in a way that causes the loader
deadlock:

    https://bugzilla.redhat.com/show_bug.cgi?id=2002656

So, if built on Fedora, we have to ship libssp, and also explicitly
link with it.  But: libssp is part of the compiler.  If we build it
ourselves from GCC source, we end up with our code built with our copy
of the import library, but the runtime code built with GCC's copy,
causing some linking to fail on missing libssp symbols.  Conversely,
the GCC Runtime Library Exception permits us to copy the compiler's DLL
into our output ZIP without complete corresponding source, but that
binary is of unknown provenance and the lack of source code looks like
a bug.

So we compromise.  We unconditionally build and ship the libssp DLL,
and only the DLL, from the GCC sources, but continue using the
compiler's import library.  That way all our binaries have consistent
linkage, and we can ship corresponding source for the DLL.  Since other
distros (Ubuntu) may not ship libssp in the MinGW runtime, we detect
whether we need -lssp to successfully link, and conditionally enable it
globally if so.  On non-Fedora systems that means we're uselessly
shipping the libssp DLL, but it's tiny.

Fixes https://github.com/openslide/openslide-winbuild/issues/23.
Fixes https://github.com/openslide/openslide-winbuild/issues/25.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>