Add a container image with build dependencies
Rather than asking users to figure out the correct package names for their
distro, provide a container image.  CI can use this too.

We'll also want to update the image on a periodic timer, but we can't do
that here because GitHub will disable the timer if the repo is inactive
for two months.  We'll let openslide/builds handle it.

Suggested in https://github.com/openslide/openslide-winbuild/issues/36#issuecomment-1174750199.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>