Build and ship winpthreads
When OpenSlide started using g_autoptr in openslide/openslide@50942fe5,
MinGW-w64 started linking it with winpthreads for pthread_once() and the
mutex functions, but only in 32-bit builds.  Apparently none of our other
dependencies use GCC cleanup functions, which is why this hasn't shown up
earlier.  It's not clear that we can avoid the dependency, and we probably
don't want to stop using g_autoptr, but libssp showed that it's possible
to build and ship our own copies of toolchain libraries.

For consistency, build and ship winpthreads in both 32-bit and 64-bit
builds.  Also stop avoiding the pthreads dependency in pixman, which means
winpthreads will provide some value on 64-bit platforms as well.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>