Skip longjmp() SEH unwind in 64-bit build
On 64-bit Windows, MinGW passes a frame pointer to _setjmp so longjmp
can do a SEH unwind.  This seems to work when the caller is also built
with MinGW, but sometimes crashes with STATUS_BAD_STACK when the caller
is built with MSVC; it appears that this is a longstanding MinGW issue.
In 64-bit builds, override setjmp() to pass a NULL frame pointer to skip
the SEH unwind.  Our uses of setjmp/longjmp are all in libpng/libjpeg
error handling, which isn't expecting to do any cleanup in intermediate
stack frames, so this should be fine.

Fixes https://github.com/openslide/openslide-winbuild/issues/47.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>