Move glib to Meson
The glib source tarball ships the gvdb source as a subproject
directory, but also has a gvdb wrap file that Meson copies into our
parent project.  Meson doesn't have a way to disable this, so ignore
the wrap file, and remove it during `clean` to transparently handle
glib version bumps.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>