Commit packagecache/CACHEDIR.TAG to repo
The tag is mainly needed when the working directory is a Git checkout,
since the package cache can accumulate source tarballs over time.  Commit
it directly rather than writing the file at runtime.

We don't copy the tag into the sdist zip.  When the working directory
comes from an sdist zip, the packagecache files won't change further, and
arguably shouldn't be excluded from archiving because they're canonically
part of the source tree at that point.

We still don't create CACHEDIR.TAG files in the unpacked subproject
directories.  Leave that to Meson to implement for now.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>