Switch builder container to Gentoo
A Gentoo-based container is significantly larger and takes much longer
to build, but gives us more control over the toolchain.  This allows us
to do three things:

- Specify the GCC and Binutils versions we want.
- Switch from msvcrt to the UCRT in both 32-bit and 64-bit builds.
- Switch the toolchain from pthreads-based thread intrinsics to
  win32-based ones so we can drop the winpthreads DLL.

The new builder container doesn't even include winpthreads, so we'll
notice if something tries to use it.  A common Meson packaging bug is to
introduce an unconditional `threads` dependency in libraries that can use
win32 threading primitives directly.

Fixes #48.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>