workflows: expect builder container repo/digest as parameter
Expect the container repo and image digest as a parameter to the windows
workflow.  This allows the caller (particularly the nightly builder) to
exercise full control over the container image, and ensures that every job
in the workflow uses the same image.

Add a find-container-digest action that can be invoked from callers'
setup jobs to compute the parameter.

This is a breaking change for callers of the windows workflow in other
repos.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>