workflows: periodically check for updates
Automatically check for updates once a week, whenever wraps are updated,
or on demand.  Have openslide-bot maintain a PR with updates from wrapdb,
and an issue for any remaining updates reported by `./build.sh updates`.

We need to do this in the context of openslide-bot, rather than the
GITHUB_TOKEN, because PRs created by the GITHUB_TOKEN can't put their
head branch in a fork, and don't automatically run CI.  However, we want
to label the issues and PRs, which would require giving openslide-bot
permissions on the repo.  Instead, we have a second, privileged workflow
to detect and label these issues/PRs.

Since the workflow maintains the PR branch in openslide-bot's fork, which
may have a stale main branch with different Actions workflows, the bot
token must have the `workflow` scope to allow updating workflows in the
branch.  Workflows have been disabled in the fork to minimize the
effects.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>