builder: add flag file noting the identity of the container image
In practice, build.sh now expects to run in the specific container image
we ship; we're not going to support other build environments.  Make that
explicit by adding a flag file that build.sh can check for.  Include a
version number, which we can increment if we substantially change the
image's functionality in future.

The build.sh side will be added in a followup after the image is rebuilt.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>