workflows: move scheduled container rebuild into this repo
The container rebuild workflow has always lived here, but the scheduled
rebuild has historically run from the openslide/builds repo.  The sole
reason for this indirection is that openslide/builds has logic to keep the
repository active, so GitHub doesn't disable scheduled jobs after 60 days
of inactivity.  Now that we have periodic update-checking in this repo, we
have that problem here anyway:

    https://github.com/openslide/openslide-winbuild/issues/144

Move the timer out of openslide/builds to simplify.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>