Generate VERSIONS.md through Meson
Meson knows what subprojects we're using in this build, which will vary
by platform.  Rather than syncing that information back into build.sh,
just generate VERSIONS.md from Meson.  For overridden subprojects that
are Git repos, get the version string from `git describe`.

Meson already has the compiler version string, so use that rather than
fetching it from the compiler.  For Binutils we still need to fetch the
version string directly.  Detect the linker command in Meson (specified in
the cross file, with a fallback to PATH) and pass it via an environment
variable so we don't need an explicit command-line argument.

Also add support for detecting the Clang version for macOS.

Give up on maintaining some sort of logical order for the dependency list,
and sort alphabetically.  The order was always fairly arbitrary.

Meson passes the MESONINTROSPECT env var to many external command
invocations, giving the correct invocation of `meson introspect`.
However, this doesn't happen for custom targets, so collect the value
via a run_command() invocation and then pass it to the custom target.

Set build_always_stale to ensure we don't package an out-of-date
VERSIONS.md.

Bump the minimum Meson version for `meson introspect --compilers` support.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>