Drop jsimdext.inc from libjpeg-turbo licenses
libjpeg-turbo's LICENSE.md stopped referring to it in 1.5.2.  Its license
text was then included in LICENSE.md until 2.0.2, but was then removed,
since the license doesn't require its text to be preserved in binary
distributions.  Binaries are the case we care about here, so we can stop
shipping the extra file.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>