Build Python wheel containing OpenSlide shared object
Generate a wheel for an openslide-bin Python package, containing the
OpenSlide shared object and an __init__.py that loads it with ctypes.
OpenSlide Python can look for this package and use its shared object
rather than requiring OpenSlide to be installed directly on the system.

Hand-roll the wheel rather than building it with an existing tool like
meson-python.  Existing Python build tools expect to be in control of the
entire build process, but we want to produce the wheel as a side effect of
an existing build.  Our needs are simple, so just implement the underlying
packaging specs directly.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>