Generate bdist archive entirely through Meson
Include support for building Linux and macOS tar.xz archives.  Change the
filename of the Windows ZIP to fully spell out the OS and CPU
architecture, and to place the version number before the platform name.
Add a machine file property specifying the platform name for use in the
archive filename.

By generating the archive inside Meson, we can avoid running
`meson install` and copying files out of the resulting directory tree.
Remove install configuration for artifacts we've generated.

Since we're not installing, we need to handle operations that Meson would
normally perform at install time, such as updating RPATHs (in
postprocess-binary.py), creating compatibility symlinks to libraries (in
write-bdist.py), and copying the import library (in
write-import-library.py).  We also need to extract the external headers
from the OpenSlide subproject, which is supported via the existing
openslide-bdist.patch.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>