Calculate OpenSlide version suffix automatically
Confusingly, we now have two suffixes: the one appended to the version
number in artifact names (`+<datestamp>.local`) and the one appended to
the OpenSlide version number (7 characters of the OpenSlide commit hash).
build.sh can calculate the latter by examining the override directory,
rather than requiring CI to pass it in, so do that to simplify.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>