bintool: don't hardcode x86_64 for Linux builds
If we use platform.machine() to get the CPU architecture, bintool can
build for Linux on any arch as long as a) it's running in a builder
container for that arch, and b) a machine file exists.

Don't add any additional containers or machine files for now, since we
can't efficiently run non-x86_64 builds in GitHub Actions.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>