Add License-File field to Python package info
To follow the source layout required by PEP 639, we need to move the
licenses directory within the wheel to *.dist-info/licenses/licenses.
We also need to add COPYING.LESSER, since it's included in the sdist.
(This is a bugfix in any case, since that license governs the artifacts
in artifacts/python.)

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>