workflows: upload source archive to PyPI as openslide_bin-*
PyPI now rejects a non-normalized package name in the source filename:

    Uploading openslide-bin-4.0.0.9.tar.gz
    WARNING  Error during upload. Retry with the --verbose option for more details.
    ERROR    HTTPError: 400 Bad Request from https://upload.pypi.org/legacy/
             Filename 'openslide-bin-4.0.0.9.tar.gz' is invalid, should be
             'openslide_bin-4.0.0.9.tar.gz'.

Rename the source tarball for PyPI but not for the GitHub release.  The
only reason we put source code on PyPI is for license compliance, since
source builds through meson-python are designed to fail.

The Source Distribution Format spec suggests that the top-level directory
name should be similarly normalized, but this isn't enforced (yet) so
ignore it for now.  We'd have to repack the output of `meson dist`, either
for PyPI or for everyone, and accept additional inconsistency in the
layout of the source archive.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>